\name{estimate}
\alias{estimate}
\title{Model Estimation}
\usage{
  estimate(object, x, ...)
}
\arguments{
  \item{object}{model to estimate}

  \item{x}{initial values for estimation}

  \item{...}{additional parameters}
}
\value{
  Estimation result
}
\description{
  Esimate a model defined by object, using x as a starting
  point.
}
\author{
  Johan Lindström
}

