% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actions.R
\name{get_action}
\alias{get_action}
\title{Get actions}
\usage{
get_action(x, action_table_name = "action", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{action_table_name}{action table name}

\item{...}{further arguments}
}
\description{
Get actions
}
\examples{

data(db)
actions <- get_action(x=db)


}
