% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Get attributes}
\usage{
get_attributes(x, attribute_table_name = "attribute", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{attribute_table_name}{attribute table name}

\item{...}{further arguments}
}
\description{
Get attributes
}
\examples{

data(db)
attributes <- get_attributes(x=db)
attributes_prcp <- get_attributes(x=db,name="PRCP_ATTRIBUTES")

}
