% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_operators.R
\name{get_operators}
\alias{get_operators}
\title{Get operators}
\usage{
get_operators(x, operator_table_name = "operator", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{operator_table_name}{operator table name}

\item{...}{further arguments}
}
\description{
Get operators
}
\examples{

data(db)
operators <- get_operators(x=db)


}
