% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p.R
\name{get_p}
\alias{get_p}
\title{Get station (or measurement points)}
\usage{
get_p(x, p = NULL, p_table_name = "p", data.table = FALSE, sf = TRUE, ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{p}{spatial (e.g. \code{sf}) object in which extract the observation station.}

\item{p_table_name}{p table name}

\item{data.table, sf, ...}{further arguments \code{\link{get_table}}}
}
\description{
Get station (or measurement points)
}
\examples{

data(db)
p <- get_p(x=db)

library(terra)
library(leaflet)

plet(vect(p),cex=10)

### IN an area 

px <- p[49,] |> st_buffer(dist=250*1000) 
p1 <- get_p(x=db,p=px)

plet(vect(p),cex=10)

px <- p[49,]
p1 <- get_p(x=db,p=px)

plet(vect(p),cex=10)


north <- -2.25
south <- -8.85
west <- 29.45
east <- 32.15
pee <- rast(xmin=west,xmax=east,ymin=south,ymax=north,crs="EPSG:4326")
p2 <- get_p(x=db,p=pee)


}
