% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(x, project_table_name = "project", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{project_table_name}{project table name}

\item{...}{further arguments}
}
\description{
Get projects
}
\examples{

data(db)
projects <- get_projects(x=db)


}
