% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_providers.R
\name{get_providers}
\alias{get_providers}
\title{Get providers}
\usage{
get_providers(x, provider_table_name = "provider", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{provider_table_name}{provider table name}

\item{...}{further arguments}
}
\description{
Get providers
}
\examples{

data(db)
providers <- get_providers(x=db)


}
