% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schema.R
\name{get_schema}
\alias{get_schema}
\title{Render/show DB schema}
\usage{
get_schema(db, ...)
}
\arguments{
\item{db}{db}

\item{...}{further arguments}
}
\description{
Render/show DB schema
}
\examples{

data(db)

library(nomnoml)



out <- get_schema(db)

nomnoml(out,png=TRUE,width=600)

}
