% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get a table, e.g. variables (or measurement types)}
\usage{
get_table(
  x,
  table_name = "ts",
  vfilter = NULL,
  data.table = (table_name != "p"),
  sf = (table_name == "p"),
  ...
)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{table_name}{variable table name}

\item{vfilter}{(possible) character string containg the filtering condition. See \code{dplyr::\link{filter}}.}

\item{data.table}{logical option. If \code{TUE} returns a \code{data.table} object}

\item{sf}{logical option.  If \code{TRUE} returns a \code{\link{sf}} object.}

\item{...}{further arguments}
}
\description{
Get a table, e.g. variables (or measurement types)
}
\examples{

data(db)
table <- get_table(x=db)
table <- get_table(x=db,table_name="ts",vfilter="p_id==4")
table2 <- get_table(x=db,table_name="ts",p_id=4)
table4 <- get_table(x=db,table_name="ts",p_id=c(3,4,21))

}
