% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variable}
\alias{get_variable}
\title{Get variables (or measurement types)}
\usage{
get_variable(x, variable_table_name = "variable", ...)
}
\arguments{
\item{x}{database or dateset, e.g  like \code{\link{db}}}

\item{variable_table_name}{variable table name}

\item{...}{further arguments}
}
\description{
Get variables (or measurement types)
}
\examples{

data(db)
variables <- get_variable(x=db)
variables <- get_variable(x=db,name="PRCP")

}
