% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdata.R
\name{print.STdata}
\alias{print.STdata}
\title{Print details for \code{STdata} object}
\usage{
\method{print}{STdata}(x, type = x$covars$type, ...)
}
\arguments{
\item{x}{\code{STdata} object to print information for.}

\item{type}{Factorial of \code{length(x$covars$ID)}, if not \code{NULL}
the output also presents summaries of number of sites and observations
as well as time periods per type of site.}

\item{...}{Ignored additional arguments.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} method for class \code{STdata}.
}
\seealso{
Other STdata methods: \code{\link{createSTdata}},
  \code{\link{plot.STdata}},
  \code{\link{print.summary.STdata}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{summary.STdata}}
}
\author{
Johan Lindstrom
}
