% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{print.summary.mcmcSTmodel}
\alias{print.summary.mcmcSTmodel}
\title{Print details for \code{summary.mcmcSTmodel} object}
\usage{
\method{print}{summary.mcmcSTmodel}(x, ...)
}
\arguments{
\item{x}{\code{summary.mcmcSTmodel} object to print information for.}

\item{...}{Additional arguments, passed to
\code{\link[base:print]{print.table}}.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} method for class \code{summary.mcmcSTmodel}.
}
\seealso{
Other mcmcSTmodel methods: \code{\link{MCMC.STmodel}},
  \code{\link{density.mcmcSTmodel}},
  \code{\link{plot.density.mcmcSTmodel}},
  \code{\link{plot.mcmcSTmodel}},
  \code{\link{print.mcmcSTmodel}},
  \code{\link{summary.mcmcSTmodel}}
}
\author{
Johan Lindstrom
}
