% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STmodel.R
\name{print.summary.STmodel}
\alias{print.summary.STmodel}
\title{Print details for \code{summary.STmodel} object}
\usage{
\method{print}{summary.STmodel}(x, ...)
}
\arguments{
\item{x}{\code{summary.STmodel} object to print information for.}

\item{...}{Ignored additional arguments.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} method for class \code{summary.STmodel}.
}
\seealso{
Other STmodel methods: \code{\link{MCMC.STmodel}},
  \code{\link{c.STmodel}}, \code{\link{createSTmodel}},
  \code{\link{estimate.STmodel}},
  \code{\link{estimateCV.STmodel}},
  \code{\link{plot.STdata}}, \code{\link{predict.STmodel}},
  \code{\link{print.STmodel}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{simulate.STmodel}},
  \code{\link{summary.STmodel}}
}
\author{
Johan Lindstrom
}
