% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_in_table.R
\name{put_in_table}
\alias{put_in_table}
\title{Put a table, e.g. variables (or measurement types)}
\usage{
put_in_table(
  x,
  db,
  table_name = "provider",
  vfilter = NULL,
  data.table = (table_name != "p"),
  sf = (table_name == "p"),
  ...
)
}
\arguments{
\item{x}{object to put in the \code{db}}

\item{db}{database or dateset, e.g  like \code{\link{db}}}

\item{table_name}{variable table name}

\item{vfilter}{(possible) character string containg the filtering condition. See \code{dplyr::\link{filter}}.}

\item{data.table}{logical option. If \code{TUE} returns a \code{data.table} object}

\item{sf}{logical option.  If \code{TRUE} returns a \code{\link{sf}} object.}

\item{...}{further arguments}
}
\description{
Put a table, e.g. variables (or measurement types)
}
\examples{

data(db)
table_name="provider"
new_providers <- data.table(name=c("asdf","qwerty"))

db <- put_in_table(new_providers,db=db,table_name=table_name)

###
table_name="variable"
new_variable_name <- "rainfall"
new_variables <- data.table(name=new_variable_name,provider_name="asdf")

db <- put_in_table(new_variables,db=db,table_name=table_name)






}
