% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STdata.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot}
\usage{
scatterPlot(x, ...)
}
\arguments{
\item{x}{object to scatter plot}

\item{...}{additional parameters}
}
\value{
Nothing
}
\description{
Scatter plot of data in x
}
\author{
Johan Lindstrom
}
