% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_matrices.R
\name{evalCovFuns}
\alias{evalCovFuns}
\title{Compute Covariance Function}
\usage{
evalCovFuns(
  type = "exp",
  pars = c(1, 1),
  d = seq(0, 10, length.out = 100),
  diff = c(0, 0)
)
}
\arguments{
\item{type}{Name of covariance functions, see \code{\link{namesCovFuns}}.}

\item{pars}{Parameter for the covariance function, see
\code{\link{parsCovFuns}}.}

\item{d}{Vector/matrix for which to compute the covariance function.}

\item{diff}{Vector with two components indicating with respect to which
parameter(s) that first and/or second derivatives should be
computed. E.g. \code{diff=c(0,0)} indicates no derivatives,
\code{diff=c(1,0)} indicates first derivative wrt the first parameter,
\code{diff=c(1,2)} indicates second cross derivative wrt the first and
second parameters, etc.}
}
\value{
Covariance function computed for all elements in d.
}
\description{
Computes covariance functions (excluding nugget) for a given vector or
matrix of distances.
}
\seealso{
Other covariance functions: 
\code{\link{crossDist}()},
\code{\link{makeSigmaB}()},
\code{\link{makeSigmaNu}()},
\code{\link{namesCovFuns}()},
\code{\link{parsCovFuns}()},
\code{\link{updateCovf}()}
}
\author{
Johan Lindstrom
}
\concept{covariance functions}
