% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{make.sigma.B}
\alias{make.sigma.B}
\alias{calc.tFX}
\alias{calc.FX}
\alias{calc.tFXF}
\alias{calc.FXtF2}
\alias{makeCholBlock}
\alias{invCholBlock}
\alias{solveTriBlock}
\alias{blockMult}
\title{Deprecated functions, use replacements!}
\usage{
make.sigma.B(...)

calc.tFX(...)

calc.FX(...)

calc.tFXF(...)

calc.FXtF2(...)

makeCholBlock(...)

invCholBlock(...)

solveTriBlock(...)

blockMult(...)
}
\arguments{
\item{...}{Unused, for compability.}
}
\value{
Does not return.
}
\description{
Deprecated functions, use replacements!
}
\details{
Functions have been rename/replaced as:
\describe{
  \item{make.sigma.B}{\code{\link{makeSigmaB}}}
  \item{calc.tFX}{Removed}
  \item{calc.FX}{Removed}
  \item{calc.tFXF}{Removed}
  \item{calc.FXtF2}{Removed}
  \item{makeCholBlock}{Removed}
  \item{invCholBlock}{Removed}
  \item{solveTriBlock}{Removed}
  \item{blockMult}{Removed}
}
}
\author{
Johan Lindstrom
}
