% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_predict.R
\name{print.predCVSTmodel}
\alias{print.predCVSTmodel}
\title{Print details for \code{predCVSTmodel} object}
\usage{
\method{print}{predCVSTmodel}(x, ...)
}
\arguments{
\item{x}{\code{predCVSTmodel} object to print information for.}

\item{...}{Ignored additional arguments.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} method for class \code{predCVSTmodel}.
}
\examples{
##load some data
data(pred.cv.mesa)
##print basic information for the CV-predictions
print(pred.cv.mesa)

}
\seealso{
Other predCVSTmodel methods: \code{\link{estimateCV.STmodel}},
  \code{\link{plot.predCVSTmodel}},
  \code{\link{print.summary.predCVSTmodel}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{summary.predCVSTmodel}}
}
\author{
Johan Lindstrom
}
