\name{print.summary.mcmcSTmodel}
\alias{print.summary.mcmcSTmodel}
\title{Print details for \code{summary.mcmcSTmodel} object}
\usage{
  \method{print}{summary.mcmcSTmodel} (x, ...)
}
\arguments{
  \item{x}{\code{summary.mcmcSTmodel} object to print
  information for.}

  \item{...}{Additional arguments, passed to
  \code{\link[base:print]{print.table}}.}
}
\value{
  Nothing
}
\description{
  \code{\link[base:print]{print}} method for class
  \code{summary.mcmcSTmodel}.
}
\author{
  Johan Lindström
}
\seealso{
  Other mcmcSTmodel methods:
  \code{\link{density.mcmcSTmodel}}, \code{\link{MCMC}},
  \code{\link{MCMC.STmodel}},
  \code{\link{plot.density.mcmcSTmodel}},
  \code{\link{plot.mcmcSTmodel}},
  \code{\link{print.mcmcSTmodel}},
  \code{\link{summary.mcmcSTmodel}}
}

