% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_predict.R
\name{print.summary.predCVSTmodel}
\alias{print.summary.predCVSTmodel}
\title{Print details for \code{summary.predCVSTmodel} object}
\usage{
\method{print}{summary.predCVSTmodel}(x, ...)
}
\arguments{
\item{x}{\code{summary.predCVSTmodel} object to print information for.}

\item{...}{Additional arguments, passed to
\code{\link[base:print]{print.table}}.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} method for class \code{summary.predCVSTmodel}.
}
\seealso{
Other predCVSTmodel methods: \code{\link{estimateCV.STmodel}},
  \code{\link{plot.predCVSTmodel}},
  \code{\link{print.predCVSTmodel}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{summary.predCVSTmodel}}
}
\author{
Johan Lindstrom
}
