\name{print.summary.predCVSTmodel}
\alias{print.summary.predCVSTmodel}
\title{Print details for \code{summary.predCVSTmodel} object}
\usage{
  \method{print}{summary.predCVSTmodel} (x, ...)
}
\arguments{
  \item{x}{\code{summary.predCVSTmodel} object to print
  information for.}

  \item{...}{Additional arguments, passed to
  \code{\link[base:print]{print.table}}.}
}
\value{
  Nothing
}
\description{
  \code{\link[base:print]{print}} method for class
  \code{summary.predCVSTmodel}.
}
\author{
  Johan Lindström
}
\seealso{
  Other predCVSTmodel methods: \code{\link{estimateCV}},
  \code{\link{estimateCV.STmodel}},
  \code{\link{plot.predCVSTmodel}},
  \code{\link{plot.predictSTmodel}},
  \code{\link{predictCV}}, \code{\link{predictCV.STmodel}},
  \code{\link{print.predCVSTmodel}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{qqnorm.STdata}},
  \code{\link{qqnorm.STmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{scatterPlot.STdata}},
  \code{\link{scatterPlot.STmodel}},
  \code{\link{summary.predCVSTmodel}}
}

