\name{print.summary.STdata}
\alias{print.summary.STdata}
\title{Print details for \code{summary.STdata} object}
\usage{
  \method{print}{summary.STdata} (x, ...)
}
\arguments{
  \item{x}{\code{summary.STdata} object to print
  information for.}

  \item{...}{Ignored additional arguments.}
}
\value{
  Nothing
}
\description{
  \code{\link[base:print]{print}} method for class
  \code{summary.STdata}.
}
\author{
  Johan Lindström
}
\seealso{
  Other STdata methods: \code{\link{createSTdata}},
  \code{\link{plot.STdata}}, \code{\link{plot.STmodel}},
  \code{\link{print.STdata}},
  \code{\link{qqnorm.predCVSTmodel}},
  \code{\link{qqnorm.STdata}},
  \code{\link{qqnorm.STmodel}},
  \code{\link{scatterPlot.predCVSTmodel}},
  \code{\link{scatterPlot.STdata}},
  \code{\link{scatterPlot.STmodel}},
  \code{\link{summary.STdata}}
}

