% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_SVD_incomplete_data.R
\name{print.SVDcv}
\alias{print.SVDcv}
\alias{summary.SVDcv}
\title{Print details for \code{SVDcv} object}
\usage{
\method{print}{SVDcv}(x, ...)

\method{summary}{SVDcv}(object, ...)
}
\arguments{
\item{x}{\code{SVDcv} object to print information for.}

\item{...}{ignored additional arguments.}

\item{object}{\code{SVDcv} object to compute summary for.}
}
\value{
Nothing
}
\description{
\code{\link[base:print]{print}} and \code{\link[base:summary]{summary}}
methods for class \code{SVDcv}, prints cross-validation statistics.
}
\examples{
  ##See SVDsmooth example

}
\seealso{
Other SVDcv methods: \code{\link{SVDsmooth}},
  \code{\link{plot.SVDcv}}

Other SVD for missing data: \code{\link{SVDmiss}},
  \code{\link{SVDsmooth}}, \code{\link{calcSmoothTrends}},
  \code{\link{plot.SVDcv}},
  \code{\link{updateTrend.STdata}}
}
\author{
Johan Lindstrom
}
